chore(engine): bulk update — refactor, fixes & build adjustments (451 files)
- Refactor shader pipeline management and resource handling
- Harden shader/pipeline loading with improved error logging
- Adjust CMake/platform initialization and related build scripts
- Code cleanup: formatting, namespace fixes, and small API clarifications
- Miscellaneous bug fixes, performance tweaks, and documentation updates

Files changed: 451

+++ kege/src/core/graphics/terrain/spherical/cubesphere/spherical-terrain-face.hpp
@@ -32,7 +32,7 @@ namespace kege{
             int depth
         );
 
-        void operator()( SphericalTerrainTile& node, const kege::vec3& position, float scale, double radius, int depth );
+        void operator()( SphericalTerrainTile& node, const kege::vec3& position, float scale, float radius, int depth );
 
         /**
          * @fn update
